requirements.txt
python-telegram-bot==20.8
replicate==0.24.0
flask==2.3.3